-Waves are no longer hard-coded (they are stored as .lvl files). I have them in the Debug folder, in a 'Game Levels' folder, which then has a folder with a specific 'level' (ex: Level0), which has 1 or more waves (ex: wave0.lvl + wave1.lvl) -Also, the gameObject class's X and Y are floats to allow slower/ more precise movement than we have been normally allowed. I have updated most speed atributes to be floats now. ~Torey

+++ trunk/CakeDefense/CakeDefense/CakeDefense/Enemy.cs
@@ -28,7 +28,7 @@ namespace CakeDefense
         #endregion Attributes
 
         #region Constructor
-        public Enemy(ImageObject imageObject, int health, int damage, int speed, Path path)
+        public Enemy(ImageObject imageObject, int health, int damage, float speed, Path path)
             : base(imageObject, health, damage, speed)
         {
             IsActive = false;
